+++ src/datasource/DronerFinityDatasource.ts
@@ -12,23 +12,21 @@ export class DronerFinityDatasource {
     provinceId?: number,
     districtId?: number,
     subdistrictId?: number,
-    creditMin?: number | null,
-    creditMax?: number | null,
+    creditMin?: number,
+    creditMax?: number,
     sortField?: string,
     sortDirection?: string,
     page?: number,
     take?: number,
   ): Promise<any> {
-    const adjustedCreditMin = creditMin === 0 ? null : creditMin
-    const adjustedCreditMax = creditMax === 0 ? null : creditMax
     const params = {
       status: status,
       search: search,
       provinceId: provinceId,
       districtId: districtId,
       subdistrictId: subdistrictId,
-      creditMin: adjustedCreditMin,
-      creditMax: adjustedCreditMax,
+      creditMin: creditMin === 0 && creditMax === 0 ? null : creditMin,
+      creditMax: creditMin === 0 && creditMax === 0 ? null : creditMax,
       sortField: sortField,
       sortDirection: sortDirection,
       page: page,
@@ -113,6 +111,18 @@ export class DronerFinityDatasource {
       .get(BASE_URL + `/droner-one-finity/find-one-droner-one-finity/${id}`)
       .then((response) => {
         return response.data.responseData
+      })
+      .catch((error) => {
+        console.log(error)
+      })
+  }
+  static downloadFile(): Promise<any> {
+    return httpClient
+      .get(BASE_URL + '/droner-one-finity/file-promise-one-finity',{
+        responseType: 'arraybuffer',
+      })
+      .then((response) => {
+        return response.data
       })
       .catch((error) => {
         console.log(error)